refactor(timeline): extract critical path logic into helper module
